Refactor: move Tenant class into its own file
Part of the ongoing breakup of binding.py
